(Asynchronous MinHashLSH) fixes critical bug when removing key from L… (#67)
(Asynchronous MinHashLSH) fixes critical bug when removing key from LSH index. Before, during removal key the buckets become broken - all similar to removal key hashes have been removed too.
