+++ docs/.buildinfo
@@ -1,4 +1,4 @@
 # Sphinx build info version 1
 # This file hashes the configuration used when building these files. When it is not found, a full rebuild will be done.
-config: 71e4b1c707b52898448eab38b1125d43
+config: 5e94d15b2373ed60f88db78c34133a53
 tags: 645f666f9bcd5a90fca523b33c5a78b7
