+++ docs/_static/basic.css
@@ -81,6 +81,10 @@ div.sphinxsidebar input {
     font-size: 1em;
 }
 
+div.sphinxsidebar #searchbox form.search {
+    overflow: hidden;
+}
+
 div.sphinxsidebar #searchbox input[type="text"] {
     float: left;
     width: 80%;
@@ -426,6 +430,13 @@ table.field-list td, table.field-list th {
     -webkit-hyphens: manual;
     hyphens: manual;
 }
+
+/* -- hlist styles ---------------------------------------------------------- */
+
+table.hlist td {
+    vertical-align: top;
+}
+
 
 /* -- other body styles ----------------------------------------------------- */
