+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["documentation","hyperloglog","index","lsh","lshensemble","lshforest","minhash","weightedminhash"],envversion:53,filenames:["documentation.rst","hyperloglog.rst","index.rst","lsh.rst","lshensemble.rst","lshforest.rst","minhash.rst","weightedminhash.rst"],objects:{"datasketch.HyperLogLog":{__eq__:[0,1,1,""],__init__:[0,1,1,""],__len__:[0,1,1,""],clear:[0,1,1,""],copy:[0,1,1,""],count:[0,1,1,""],digest:[0,1,1,""],is_empty:[0,1,1,""],merge:[0,1,1,""],update:[0,1,1,""]},"datasketch.HyperLogLogPlusPlus":{__weakref__:[0,2,1,""],count:[0,1,1,""]},"datasketch.LeanMinHash":{__hash__:[0,1,1,""],__init__:[0,1,1,""],bytesize:[0,1,1,""],copy:[0,1,1,""],deserialize:[0,3,1,""],serialize:[0,1,1,""],union:[0,3,1,""],update:[0,1,1,""]},"datasketch.MinHash":{__eq__:[0,1,1,""],__init__:[0,1,1,""],__len__:[0,1,1,""],__weakref__:[0,2,1,""],clear:[0,1,1,""],copy:[0,1,1,""],count:[0,1,1,""],digest:[0,1,1,""],is_empty:[0,1,1,""],jaccard:[0,1,1,""],merge:[0,1,1,""],union:[0,3,1,""],update:[0,1,1,""]},"datasketch.MinHashLSH":{__contains__:[0,1,1,""],__init__:[0,1,1,""],__weakref__:[0,2,1,""],get_counts:[0,1,1,""],get_subset_counts:[0,1,1,""],insert:[0,1,1,""],insertion_session:[0,1,1,""],is_empty:[0,1,1,""],query:[0,1,1,""],remove:[0,1,1,""]},"datasketch.MinHashLSHEnsemble":{__contains__:[0,1,1,""],__init__:[0,1,1,""],__weakref__:[0,2,1,""],index:[0,1,1,""],is_empty:[0,1,1,""],query:[0,1,1,""]},"datasketch.MinHashLSHForest":{__contains__:[0,1,1,""],__init__:[0,1,1,""],__weakref__:[0,2,1,""],add:[0,1,1,""],index:[0,1,1,""],is_empty:[0,1,1,""],query:[0,1,1,""]},"datasketch.WeightedMinHash":{__eq__:[0,1,1,""],__init__:[0,1,1,""],__len__:[0,1,1,""],__weakref__:[0,2,1,""],copy:[0,1,1,""],digest:[0,1,1,""],jaccard:[0,1,1,""]},"datasketch.WeightedMinHashGenerator":{__init__:[0,1,1,""],__weakref__:[0,2,1,""],minhash:[0,1,1,""]},"datasketch.experimental.aio.lsh":{AsyncMinHashLSH:[0,0,1,""]},"datasketch.experimental.aio.lsh.AsyncMinHashLSH":{__init__:[0,1,1,""],__weakref__:[0,2,1,""],close:[0,1,1,""],get_counts:[0,1,1,""],get_subset_counts:[0,1,1,""],has_key:[0,1,1,""],insert:[0,1,1,""],insertion_session:[0,1,1,""],is_empty:[0,1,1,""],query:[0,1,1,""],remove:[0,1,1,""]},datasketch:{HyperLogLog:[0,0,1,""],HyperLogLogPlusPlus:[0,0,1,""],LeanMinHash:[0,0,1,""],MinHash:[0,0,1,""],MinHashLSH:[0,0,1,""],MinHashLSHEnsemble:[0,0,1,""],MinHashLSHForest:[0,0,1,""],WeightedMinHash:[0,0,1,""],WeightedMinHashGenerator:[0,0,1,""]}},objnames:{"0":["py","class","Python class"],"1":["py","method","Python method"],"2":["py","attribute","Python attribute"],"3":["py","classmethod","Python class method"]},objtypes:{"0":"py:class","1":"py:method","2":"py:attribute","3":"py:classmethod"},terms:{"\u00e9ric":1,"break":[],"byte":0,"case":[4,5],"class":0,"default":[0,1,5,6,7],"export":0,"float":[0,6],"fr\u00e9d\u00e9ric":1,"function":[0,1,3,4,5,6,7],"import":[0,1,3,4,5,6,7],"int":[0,5],"long":7,"new":[0,3],"return":[0,1,3,5,6],"super":2,"true":0,"try":0,AND:0,But:4,For:[0,2,3],One:0,The:[0,1,2,3,4,5,6],There:[3,4,5],Use:[0,1],Used:0,Useful:[],Using:[0,3,4,5,7],__contains__:0,__eq__:0,__hash__:0,__init__:0,__len__:0,__weakref__:0,_async_minhash_lsh_at_scal:[],_batch_siz:[],_chunked_str:[0,3],_minhash_lsh_async:[],_minhash_lsh_at_scal:[],_storag:3,_storage_config_redi:0,aac:[0,3],aahh:[0,3],aahhb:[0,3],aahhc:[0,3],about:[0,3,4],abov:[0,2,3,4],accur:0,accuraci:[0,1,2,3,4,6,7],achiev:[3,4],across:[0,3],actual:[1,6],actual_jaccard:6,add:[0,5],added:0,adding:5,addit:[0,3],adjust:[0,6],advis:3,affect:0,after:[0,5],aha:[],aio:0,aiohttp:[],aioinsert_syncredis_with_executor:[],aiomongo:[0,3],aioquery_syncredi:[],aioredi:[0,3],algorithm:[0,3],all:[0,1,3,6,7],alloc:0,allow:3,also:[0,1,2,3,4,6,7],altern:3,alwai:3,amount:2,analysi:6,andrei:6,ani:0,anoth:[0,4,5],api:3,approach:[3,7],approxim:[0,3,7],arbitrari:[4,6],arg:[],argument:[0,3,7],arrai:0,ascii_lowercas:[0,3],assign:0,associ:3,assum:[0,7],assur:3,async:3,async_redis_storage_benchmark:[],asynchron:2,asyncio:[0,3],asyncminhash:[],asyncminhashlsh:[0,3],asyncminhashlshinsertionsess:0,asyncmongostorag:[],asyncredisstorag:[],avail:0,averag:[3,5],avoid:0,await:[0,3],awaitebl:[],band:0,base:[0,1,3],base_nam:[],base_name_1:0,basenam:[0,3],batch:3,batch_siz:[0,3],bawa:5,becaus:[4,6],becom:4,been:0,befor:0,being:4,below:[3,4,5],benchmark:[3,4,5],best:3,better:[0,1,4,6,7],between:[0,3,4,5,6],bhg:[0,3],bia:1,bias:4,bigger:3,bike:4,billion:6,binari:7,bird:4,bit:[0,1],block:3,bool:0,both:[0,6],broder:6,bucket:0,buf:0,buffer:0,buffer_s:[0,3],built:0,bulk:3,bypass:0,bytearrai:0,byteord:0,bytes:0,call:[0,3,4,5],can:[0,1,2,3,4,6,7],candid:5,cannot:[0,3,7],capabl:1,car:4,cardin:[0,1,2,3,6],cat:4,certain:[3,4],chain:[0,3],chang:[0,1,3],chapter:[0,3],charact:0,check:[0,4,5],check_dupl:[0,3],choic:[0,3],chose:[],chosen:0,chunk:[0,3],classic:3,classmethod:0,cleanup:0,clear:[0,4],client:0,close:[0,3],cluster:3,code:[3,4],cohen:6,coher:5,col:[],collect:[3,4],collis:[0,3],com:[0,4],come:3,comment:[],compress:7,comput:[0,3,4,6],concurr:3,config:3,configur:[0,3],connect:[],consid:[0,3,6],consist:0,constant:4,constrain:[],constructor:0,contain:[0,2],context:[0,3],control:[0,1],convert:0,copi:0,correct:1,correspond:0,cost:[3,7],could:[],count:[0,1,6,7],cow:4,cpu:6,creat:[0,1,3,4,5,7],current:[0,3],custom:[5,6],data1:[1,5,6],data2:[5,6],data3:5,data:[0,1,3,4,5,6,7],data_list:3,dataset:[0,1,3,5,6],datasketch:[0,1,3,4,5,6,7],decreas:4,deduc:4,def:[0,3],defin:0,denomin:4,depend:[2,5,7],depth:5,describ:0,deseri:[0,3],detail:[3,4,5],determin:4,diagram:4,dict:0,differ:[0,3,4,5],digest:0,dim:0,dimens:[0,7],directli:4,directori:[3,4],disconnect:0,distinct:[1,6],divid:4,document:[3,4,5,6],doe:[0,3,7],dog:4,doing:7,don:3,doubl:[],due:[0,3],dump:3,duplic:0,dure:[0,3],each:[0,7],easili:4,eer:[0,3],effect:0,effici:3,ekzhu:4,elaps:[],element:[0,6],eleph:4,els:1,empti:0,encod:[0,1,3,4,5,6],endpoint:[],enhanc:[0,1],enough:[],ensembl:[2,3],ensur:0,entri:0,enumer:0,environ:3,equal:0,equival:0,estim:[0,1,2,3,5,6,7],etc:[],everi:[0,3,5,6],everyth:[1,3],exact:3,exampl:[0,3,7],except:[],execut:[],exist:0,expand:7,expect:3,expens:[0,7],experi:[1,3,4],experiment:[0,2,3],exponenti:1,extern:3,extrem:3,face:[],factor:0,fals:[0,3],false_negative_weight:0,false_positive_weight:0,far:[0,6],fast:[0,2],faster:[0,2,3],find:[3,4],first:[0,1,3,7],fish:4,fix:[1,4,5,6],flajolet:1,follow:[1,2,4],footprint:[0,6],forest:[2,3],fork:3,form:0,format:0,found:[0,3,4],fraction:4,frequenc:0,from:[0,1,3,4,5,6,7],frozen:0,frozenset:[0,3],full:4,further:0,fusi:1,futur:[0,3],gandouet:1,gather:[0,3],gener:[0,3,5],get:[0,3,4,7],get_count:0,get_subset_count:0,github:[0,4],give:[0,1,2,3,6],given:[0,4],googl:[0,1],great:4,greater:0,group:3,guarante:[],handl:[3,6],has:[0,3],has_kei:0,hash:[0,1,3,6],hashabl:0,hashlib:0,hashobj:0,hashtabl:0,hashvalu:0,have:[0,2,3,4,5,7],header:[],hei:[],help:0,here:[0,1,4,7],higest:5,high:[0,3],higher:[0,1,3,6],highest:[0,5],hors:4,host:[0,3],howev:[1,3,4,5,7],hpp:1,http:0,huge:3,human:4,hurri:[],hyperloglog:2,hyperloglogplusplu:[0,1],identifi:[0,3],ignor:0,illustr:4,implement:[0,1,3,4,5,7],improv:[0,3],includ:[3,5],inconsist:3,increas:[3,4,7],index:[0,2,3,4,5,7],inform:0,inherit:0,initi:[0,1,2,3],input:0,insert:[0,3,7],insert_aioredi:[],insert_sess:0,insertion_sess:[0,3],insertion_session_aioredi:[],insertion_session_syncredi:[],instead:[0,1,4],instruct:6,integ:0,intent:3,interestingli:1,interfac:[0,1,3,5,7],intern:0,intersect:[3,4,6],introduc:[1,6],ioff:[0,7],is_empti:0,islic:[0,3],issu:4,item:[0,7],iter:[0,3,4],its:[0,4,7],jaccard:[0,2,3,4,5,6,7],jet:4,join:[0,3],jump:3,just:[0,4],kei:[0,3,4,5],kkd:[0,3],kld:[0,3],know:3,knowledg:7,known:7,lambda:[0,3],languag:0,larg:[0,2,3,4,7],larger:4,latter:4,layer:[2,3],lead:0,lean_minhash2:0,lean_minhash:0,leanminhash:[0,6],least:0,left:4,len:[0,1,3,4,6,7],length:0,let:6,librari:7,like:[3,4],linear:[2,3,5,6],linearli:[3,6,7],lion:4,list:0,littl:2,live:0,lmh:0,load:3,local:3,localhost:[0,3],locat:3,loss:2,lower:3,lsh:[2,7],lsh_test:3,lshensembl:4,mai:[0,3,5],main:0,maintain:0,make:[0,1,3,5,6,7],manag:[0,3],mani:[1,3,5],map:5,mapreduc:6,massiv:[0,3],match:5,max_work:3,maxim:3,maximum:[0,5],maxk:0,mean:[3,5,6],measur:[3,4],membership:[4,5],memori:[0,1,6],merg:[0,1,6],method:[0,5],meunier:1,mhs:0,migrat:0,million:3,mine:[0,3],minhash2:0,minhash:[1,2],minhash_lsh_at_scal:[],minhashlsh:[0,3,5,7],minhashlshensembl:[0,4],minhashlshforest:[0,5,7],minhashlshinsertionsess:0,minim:0,miniz:0,mode:[0,3],modif:0,modifi:7,modul:[0,2,3],mongo:[0,3],mongodb:[2,3],more:[0,1,3,4,5,6],moreov:3,most:0,motor:3,motorcyl:4,much:[2,3,4],multi:0,multipl:[0,3],multipli:4,multiset:7,must:[0,2,3,5],name:[0,3],nativ:0,need:[0,3],neg:[0,3],neighbour:[3,7],network:3,newer:0,next:0,none:0,normal:4,note:[0,2,3],now:[0,3],num_part:0,num_perm:[0,3,4,5,6],number:[0,1,3,4,5,6,7],numpi:[0,2,7],obj:[0,3],object:[0,1,3,4,5,6,7],obvious:[],off:6,often:5,old:3,olivi:1,onc:0,one:[0,4],onli:[0,7],open:3,openssl_sha1:0,optim:[0,3],option:[0,2,3,4,5],order:0,origin:[0,1,6],other:[0,3,4],otherwis:5,our:[],over:0,overhead:0,packag:[0,2,3,4],pair:4,paper:[0,1,5,6],parallel:[0,6],param:0,paramet:[0,1,3,4,5,7],part:3,partit:0,pass:[0,1,3],penal:4,penalti:1,percentil:3,perform:[0,4,7],permut:[0,3,4,6,7],persist:3,philipp:1,pickl:[0,3],pig:4,pip:2,plane:4,plot:[3,4,5],pool:3,popular:3,port:[0,3],posit:[0,3],possibl:[5,7],ppi:[0,3],precis:[0,3,5],prefer:0,prefix:[0,3,5],prepickl:0,present:6,preserv:3,prevent:0,print:[1,3,4,5,6,7],probabilist:[0,1,2,3,5,6],probabl:[3,5,6],problem:4,process:[2,3,7],product:3,profil:7,program:0,properli:0,properti:[0,3],proport:6,provid:[2,3],pymongo:3,python:[0,2,3],qualifi:3,quer:[],queri:[0,2,3,4,5,7],query_aioredi:[],query_syncredi:[],radiu:5,rais:0,ram:[],random:[0,3,7],rang:[0,3],read:5,real:7,recal:[0,3],recommend:[0,3],redi:[0,2,3],reduc:[3,6],ref:[],refer:0,referenc:0,reg:0,rel:0,remov:0,repositori:[3,4],repres:0,represent:[0,1],requir:[3,7],resembl:[0,4,6],reset:0,resourc:0,respect:[0,3,4,7],respons:0,rest:[0,3],restrict:0,result:[0,3,4,5,7],retriev:[0,5],right:[3,4],rocket:4,roughli:4,run_in_executor:[],said:3,same:[0,1,3,4,5],sampl:[0,7],sample_s:[0,7],satisfi:3,save:[0,3],scale:0,scan:5,schema:0,scheme:[0,1],scipi:2,scooter:4,search:[2,3,4],searchabl:[0,5],searchbal:0,sec:3,see:[0,3,4,5],seed:[0,7],seen:[0,6],self:0,sensit:3,separ:[],seq:[0,3],sequenc:3,sergei:[0,7],serial:[0,3],serializ:3,server:[],servic:[0,3],session:[0,3],set1:[3,4],set2:[3,4],set3:[3,4],set:[0,1,3,4,5,6,7],sha1:0,share:[1,3],shingl:3,ship:4,should:[],show:[4,5],side:0,signatur:0,similar:[0,2,3,4,5,6,7],simpl:[],simpli:3,simplifi:4,sinc:[0,3,4],singl:[0,1,3],size:[0,3,4,6],sketch:[0,2,4,5],slightli:[0,3,4],slower:[0,3,6,7],small:[0,1,6],smaller:[0,4],sole:4,some:5,soon:4,sort:0,sourc:[0,3,4],space:[0,1,6],spars:[0,1],specif:4,specifi:[0,3,7],speed:[1,6,7],spread:[],stabl:1,start:[],state:0,step:0,still:3,storag:[0,2,3],storage_config:[0,3],store:[0,6],str:0,stream:[3,6,7],string:[0,3],structur:[0,1,2,3,5,6],style:[3,6],sub:[2,3],submarin:4,sum:0,summari:3,support:[0,2,3,7],suppos:[3,4],suv:4,svpcom:0,synchron:3,synthet:5,tabl:0,take:[4,5,7],task:3,techniqu:0,test:[1,3],text:3,than:[0,1,3,6],them:0,ther:3,thi:[0,1,2,3,4,5,6,7],think:4,those:0,though:3,threadpoolexecutor:3,threshold:[0,2,3,4,5,7],through:[],throughput:3,thu:[3,4,7],tiger:4,time:[0,2,3,6],togeth:0,top:[0,2,3,5],toward:0,trade:6,train:4,tree:[0,5],tune:[3,4],tupl:[0,3],twice:[],two:[0,1,3,4,6],type:[0,2,3,5,7],typeerror:0,uderstand:[],under:0,underli:3,understand:5,uniform:7,union:[0,1,4,6],uniqu:[0,3],unique_name_6ac4fg:3,unit:0,univers:7,unlik:3,unseen:7,unsuit:7,until:0,unweight:7,updat:[0,1,3,4,5,6,7],usag:[0,1,2,6,7],use:[0,1,3,4,7],used:[0,1,2,3,4,5,6,7],useful:[3,5,6],uses:0,using:[0,1,2,3,4,5,6,7],usual:[3,7],utf8:[1,3,4,5,6],utf:[0,3],valu:[0,1,4,6],van:4,variabl:0,variou:3,vector:[0,7],venn:4,veri:[0,2,3,4,5,7],version:[0,1,3,4],wai:[3,4,5],want:[0,3,4,5],weak:0,web:[],weight:2,weightedminhash:[0,7],weightedminhashgener:[0,7],when:[0,3,4,7],where:0,whether:0,which:[0,3,4,5,7],who:[],whose:0,why:4,without:[0,5],wm1:7,wm2:7,wmg:7,wolf:4,won:[0,5],work:[0,4],worker:[],would:[],wouldn:[],yes:[],you:[0,1,2,3,4,5,6,7],your:[3,4,6],yow:[0,3],zhu:[0,4],zip:[0,3]},titles:["API Documentation","HyperLogLog","datasketch: Big Data Looks Small","MinHash LSH","MinHash LSH Ensemble","MinHash LSH Forest","MinHash","Weighted MinHash"],titleterms:{api:0,asynchron:[0,3],big:2,connect:3,contain:4,data:2,datasketch:2,document:0,ensembl:[0,4],exist:3,forest:[0,5],hyperloglog:[0,1],instal:2,lean:0,look:2,lsh:[0,3,4,5],minhash:[0,3,4,5,6,7],scale:3,small:2,weight:[0,7]}})
+Search.setIndex({docnames:["documentation","hyperloglog","index","lsh","lshensemble","lshforest","minhash","weightedminhash"],envversion:{"sphinx.domains.c":1,"sphinx.domains.changeset":1,"sphinx.domains.cpp":1,"sphinx.domains.javascript":1,"sphinx.domains.math":1,"sphinx.domains.python":1,"sphinx.domains.rst":1,"sphinx.domains.std":1,"sphinx.ext.viewcode":1,sphinx:55},filenames:["documentation.rst","hyperloglog.rst","index.rst","lsh.rst","lshensemble.rst","lshforest.rst","minhash.rst","weightedminhash.rst"],objects:{"datasketch.HyperLogLog":{__eq__:[0,1,1,""],__init__:[0,1,1,""],__len__:[0,1,1,""],clear:[0,1,1,""],copy:[0,1,1,""],count:[0,1,1,""],digest:[0,1,1,""],is_empty:[0,1,1,""],merge:[0,1,1,""],update:[0,1,1,""]},"datasketch.HyperLogLogPlusPlus":{__weakref__:[0,2,1,""],count:[0,1,1,""]},"datasketch.LeanMinHash":{__hash__:[0,1,1,""],__init__:[0,1,1,""],bytesize:[0,1,1,""],copy:[0,1,1,""],deserialize:[0,3,1,""],serialize:[0,1,1,""],union:[0,3,1,""],update:[0,1,1,""]},"datasketch.MinHash":{__eq__:[0,1,1,""],__init__:[0,1,1,""],__len__:[0,1,1,""],__weakref__:[0,2,1,""],clear:[0,1,1,""],copy:[0,1,1,""],count:[0,1,1,""],digest:[0,1,1,""],is_empty:[0,1,1,""],jaccard:[0,1,1,""],merge:[0,1,1,""],union:[0,3,1,""],update:[0,1,1,""]},"datasketch.MinHashLSH":{__contains__:[0,1,1,""],__init__:[0,1,1,""],__weakref__:[0,2,1,""],get_counts:[0,1,1,""],get_subset_counts:[0,1,1,""],insert:[0,1,1,""],insertion_session:[0,1,1,""],is_empty:[0,1,1,""],query:[0,1,1,""],remove:[0,1,1,""]},"datasketch.MinHashLSHEnsemble":{__contains__:[0,1,1,""],__init__:[0,1,1,""],__weakref__:[0,2,1,""],index:[0,1,1,""],is_empty:[0,1,1,""],query:[0,1,1,""]},"datasketch.MinHashLSHForest":{__contains__:[0,1,1,""],__init__:[0,1,1,""],__weakref__:[0,2,1,""],add:[0,1,1,""],index:[0,1,1,""],is_empty:[0,1,1,""],query:[0,1,1,""]},"datasketch.WeightedMinHash":{__eq__:[0,1,1,""],__init__:[0,1,1,""],__len__:[0,1,1,""],__weakref__:[0,2,1,""],copy:[0,1,1,""],digest:[0,1,1,""],jaccard:[0,1,1,""]},"datasketch.WeightedMinHashGenerator":{__init__:[0,1,1,""],__weakref__:[0,2,1,""],minhash:[0,1,1,""]},"datasketch.experimental.aio.lsh":{AsyncMinHashLSH:[0,0,1,""]},"datasketch.experimental.aio.lsh.AsyncMinHashLSH":{__init__:[0,1,1,""],__weakref__:[0,2,1,""],close:[0,1,1,""],get_counts:[0,1,1,""],get_subset_counts:[0,1,1,""],has_key:[0,1,1,""],insert:[0,1,1,""],insertion_session:[0,1,1,""],is_empty:[0,1,1,""],query:[0,1,1,""],remove:[0,1,1,""]},datasketch:{HyperLogLog:[0,0,1,""],HyperLogLogPlusPlus:[0,0,1,""],LeanMinHash:[0,0,1,""],MinHash:[0,0,1,""],MinHashLSH:[0,0,1,""],MinHashLSHEnsemble:[0,0,1,""],MinHashLSHForest:[0,0,1,""],WeightedMinHash:[0,0,1,""],WeightedMinHashGenerator:[0,0,1,""]}},objnames:{"0":["py","class","Python class"],"1":["py","method","Python method"],"2":["py","attribute","Python attribute"],"3":["py","classmethod","Python class method"]},objtypes:{"0":"py:class","1":"py:method","2":"py:attribute","3":"py:classmethod"},terms:{"\u00e9ric":1,"byte":0,"case":[4,5],"class":0,"default":[0,1,5,6,7],"export":0,"float":[0,6],"fr\u00e9d\u00e9ric":1,"function":[0,1,3,4,5,6,7],"import":[0,1,3,4,5,6,7],"int":[0,5],"long":7,"new":[0,3],"return":[0,1,3,5,6],"super":2,"true":0,"try":0,AND:0,But:4,For:[0,2,3],One:0,The:[0,1,2,3,4,5,6],There:[3,4,5],Use:[0,1],Used:0,Using:[0,3,4,5,7],__contains__:0,__eq__:0,__hash__:0,__init__:0,__len__:0,__weakref__:0,_chunked_str:[0,3],_storag:3,_storage_config_redi:0,aac:[0,3],aahh:[0,3],aahhb:[0,3],aahhc:[0,3],about:[0,3,4],abov:[0,2,3,4],accur:0,accuraci:[0,1,2,3,4,6,7],achiev:[3,4],across:[0,3],actual:[1,6],actual_jaccard:6,add:[0,5],added:0,adding:5,addit:[0,3],adjust:[0,6],advis:3,affect:0,after:[0,5],aio:[0,3],aiomongo:[0,3],aioredi:[0,3],algorithm:[0,3],all:[0,1,3,6,7],alloc:0,allow:3,also:[0,1,2,3,4,6,7],altern:3,alwai:3,amount:2,analysi:6,andrei:6,ani:0,anoth:[0,4,5],api:3,approach:[3,7],approxim:[0,3,7],arbitrari:[4,6],argument:[0,3,7],arrai:0,ascii_lowercas:[0,3],assign:0,associ:3,assum:[0,7],assur:3,async:[0,3],asynchron:2,asyncio:[0,3],asyncminhashlsh:[0,3],asyncminhashlshinsertionsess:0,avail:0,averag:[3,5],avoid:0,await:[0,3],band:0,base:[0,1,3],base_name_1:0,basenam:[0,3],batch:3,batch_siz:[0,3],bawa:5,becaus:[4,6],becom:4,been:0,befor:0,being:4,below:[3,4,5],benchmark:[3,4,5],best:3,better:[0,1,4,6,7],between:[0,3,4,5,6],bhg:[0,3],bia:1,bias:4,bigger:3,bike:4,billion:6,binari:7,bird:4,bit:[0,1],block:3,bool:0,both:[0,6],broder:6,bucket:0,buf:0,buffer:0,buffer_s:[0,3],built:0,bulk:3,bypass:0,bytearrai:0,byteord:0,bytes:0,call:[0,3,4,5],can:[0,1,2,3,4,6,7],candid:5,cannot:[0,3,7],cap:4,capabl:1,car:4,cardin:[0,1,2,3,6],cat:4,certain:[3,4],chain:[0,3],chang:[0,1,3],chapter:[0,3],charact:0,check:[0,4,5],check_dupl:[0,3],choic:[0,3],chosen:0,chunk:[0,3],classic:3,classmethod:0,cleanup:0,clear:[0,4],client:0,close:[0,3],cluster:3,code:[3,4],cohen:6,coher:5,collect:[3,4],collis:[0,3],com:[0,4],come:3,compress:7,comput:[0,3,4,6],concurr:3,config:3,configur:[0,3],consid:[0,3,6],consist:0,constant:4,constructor:0,contain:[0,2],context:[0,3],control:[0,1],convert:0,copi:0,correct:1,correspond:0,cost:[3,7],count:[0,1,6,7],cow:4,cpu:6,creat:[0,1,3,4,5,7],cup:4,current:[0,3],custom:[5,6],data1:[1,5,6],data2:[5,6],data3:5,data:[0,1,3,4,5,6,7],data_list:3,dataset:[0,1,3,5,6],datasketch:[0,1,3,4,5,6,7],decreas:4,deduc:4,def:[0,3],defin:0,denomin:4,depend:[2,5,7],depth:5,describ:0,deseri:[0,3],detail:[3,4,5],determin:4,diagram:4,dict:0,differ:[0,3,4,5],digest:0,dim:0,dimens:[0,7],directli:4,directori:[3,4],disconnect:0,distinct:[1,6],divid:4,document:[3,4,5,6],doe:[0,3,7],dog:4,doing:7,don:3,due:[0,3],dump:3,duplic:0,dure:[0,3],each:[0,7],easili:4,eer:[0,3],effect:0,effici:3,ekzhu:4,element:[0,6],eleph:4,els:1,empti:0,encod:[0,1,3,4,5,6],enhanc:[0,1],ensembl:[2,3],ensur:0,entri:0,enumer:0,environ:3,equal:0,equival:0,estim:[0,1,2,3,5,6,7],everi:[0,3,5,6],everyth:[1,3],exact:3,exampl:[0,3,7],exist:0,expand:7,expect:3,expens:[0,7],experi:[1,3,4],experiment:[0,2,3],exponenti:1,extern:3,extrem:3,factor:0,fals:[0,3],false_negative_weight:0,false_positive_weight:0,far:[0,6],fast:[0,2],faster:[0,2,3],find:[3,4],first:[0,1,3,7],fish:4,fix:[1,4,5,6],flajolet:1,follow:[1,2,4],footprint:[0,6],forest:[2,3],fork:3,form:0,format:0,found:[0,3,4],frac:4,fraction:4,frequenc:0,from:[0,1,3,4,5,6,7],frozen:0,frozenset:[0,3],full:4,func:[0,3],further:0,fusi:1,futur:[0,3],gandouet:1,gather:[0,3],gener:[0,3,5],get:[0,3,4,7],get_count:0,get_subset_count:0,github:[0,4],give:[0,1,2,3,6],given:[0,4],googl:[0,1],great:4,greater:0,group:3,handl:[3,6],has:[0,3],has_kei:0,hash:[0,1,3,6],hashabl:0,hashlib:0,hashobj:0,hashtabl:0,hashvalu:0,have:[0,2,3,4,5,7],help:0,here:[0,1,4,7],higest:5,high:[0,3],higher:[0,1,3,6],highest:[0,5],hors:4,host:[0,3],howev:[1,3,4,5,7],hpp:1,http:0,huge:3,human:4,hyperloglog:2,hyperloglogplusplu:[0,1],identifi:[0,3],ignor:0,illustr:4,implement:[0,1,3,4,5,7],improv:[0,3],includ:[3,5],inconsist:3,increas:[3,4,7],index:[0,2,3,4,5,7],inform:0,inherit:0,initi:[0,1,2,3],input:0,insert:[0,3,7],insert_sess:0,insertion_sess:[0,3],instead:[0,1,4],instruct:6,integ:0,intent:3,interestingli:1,interfac:[0,1,3,5,7],intern:0,intersect:[3,4,6],introduc:[1,6],ioff:[0,7],is_empti:0,islic:[0,3],issu:4,item:[0,7],iter:[0,3,4],its:[0,4,7],jaccard:[0,2,3,4,5,6,7],jet:4,join:[0,3],jump:3,just:[0,4],kei:[0,3,4,5],kkd:[0,3],kld:[0,3],know:3,knowledg:7,known:7,lambda:[0,3],languag:0,larg:[0,2,3,4,7],larger:4,latter:4,layer:[2,3],lead:0,lean_minhash2:0,lean_minhash:0,leanminhash:[0,6],least:0,left:4,len:[0,1,3,4,6,7],length:0,let:6,librari:7,like:[3,4],linear:[2,3,5,6],linearli:[3,6,7],lion:4,list:0,littl:2,live:0,lmh:0,load:3,local:3,localhost:[0,3],locat:3,loss:2,lower:3,lsh:[2,7],lsh_test:3,lshensembl:4,mai:[0,3,5],main:0,maintain:0,make:[0,1,3,5,6,7],manag:[0,3],mani:[1,3,5],map:5,mapreduc:6,massiv:[0,3],match:5,max_work:3,maxim:3,maximum:[0,5],maxk:0,mean:[3,5,6],measur:[3,4],membership:[4,5],memori:[0,1,6],merg:[0,1,6],method:[0,5],meunier:1,mhs:0,migrat:0,million:3,mine:[0,3],minhash2:0,minhash:[1,2],minhashlsh:[0,3,5,7],minhashlshensembl:[0,4],minhashlshforest:[0,5,7],minhashlshinsertionsess:0,minim:0,miniz:0,mode:[0,3],modif:0,modifi:7,modul:[0,2,3],mongo:[0,3],mongodb:[2,3],more:[0,1,3,4,5,6],moreov:3,most:0,motor:3,motorcyl:4,much:[2,3,4],multi:0,multipl:[0,3],multipli:4,multiset:7,must:[0,2,3,5],name:[0,3],nativ:0,need:[0,3],neg:[0,3],neighbour:[3,7],network:3,newer:0,next:0,none:0,normal:4,note:[0,2,3],now:[0,3],num_part:0,num_perm:[0,3,4,5,6],number:[0,1,3,4,5,6,7],numpi:[0,2,7],obj:[0,3],object:[0,1,3,4,5,6,7],off:6,often:5,old:3,olivi:1,onc:0,one:[0,4],onli:[0,7],open:3,openssl_sha1:0,optim:[0,3],option:[0,2,3,4,5],order:0,origin:[0,1,6],other:[0,3,4],otherwis:5,over:0,overhead:0,packag:[0,2,3,4],pair:4,paper:[0,1,5,6],parallel:[0,6],param:0,paramet:[0,1,3,4,5,7],part:3,partit:0,pass:[0,1,3],penal:4,penalti:1,percentil:3,perform:[0,4,7],permut:[0,3,4,6,7],persist:3,philipp:1,pickl:[0,3],pig:4,pip:2,plane:4,plot:[3,4,5],pool:3,popular:3,port:[0,3],posit:[0,3],possibl:[5,7],ppi:[0,3],precis:[0,3,5],prefer:0,prefix:[0,3,5],prepickl:0,present:6,preserv:3,prevent:0,print:[1,3,4,5,6,7],probabilist:[0,1,2,3,5,6],probabl:[3,5,6],problem:4,process:[2,3,7],product:3,profil:7,program:0,properli:0,properti:[0,3],proport:6,provid:[2,3],pymongo:3,python:[0,2,3],qualifi:3,queri:[0,2,3,4,5,7],radiu:5,rais:0,random:[0,3,7],rang:[0,3],read:5,real:7,recal:[0,3],recommend:[0,3],redi:[0,2,3],reduc:[3,6],refer:0,referenc:0,reg:0,rel:0,remov:0,repositori:[3,4],repres:0,represent:[0,1],requir:[3,7],resembl:[0,4,6],reset:0,resourc:0,respect:[0,3,4,7],respons:0,rest:[0,3],restrict:0,result:[0,3,4,5,7],retriev:[0,5],right:[3,4],rocket:4,roughli:4,said:3,same:[0,1,3,4,5],sampl:[0,7],sample_s:[0,7],satisfi:3,save:[0,3],scale:0,scan:5,schema:0,scheme:[0,1],scipi:2,scooter:4,search:[2,3,4],searchabl:[0,5],searchbal:0,sec:3,see:[0,3,4,5],seed:[0,7],seen:[0,6],self:0,sensit:3,seq:[0,3],sequenc:3,sergei:[0,7],serial:[0,3],serializ:3,servic:[0,3],session:[0,3],set1:[3,4],set2:[3,4],set3:[3,4],set:[0,1,3,4,5,6,7],sha1:0,share:[1,3],shingl:3,ship:4,show:[4,5],side:0,signatur:0,similar:[0,2,3,4,5,6,7],simpli:3,simplifi:4,sinc:[0,3,4],singl:[0,1,3],size:[0,3,4,6],sketch:[0,2,4,5],slightli:[0,3,4],slower:[0,3,6,7],small:[0,1,6],smaller:[0,4],sole:4,some:5,soon:4,sort:0,sourc:[0,3,4],space:[0,1,6],spars:[0,1],specif:4,specifi:[0,3,7],speed:[1,6,7],stabl:1,state:0,step:0,still:3,storag:[0,2,3],storage_config:[0,3],store:[0,6],str:0,stream:[3,6,7],string:[0,3],structur:[0,1,2,3,5,6],style:[3,6],sub:[2,3],submarin:4,sum:0,summari:3,support:[0,2,3,7],suppos:[3,4],suv:4,svpcom:0,synchron:3,synthet:5,tabl:0,take:[4,5,7],task:3,techniqu:0,test:[1,3],text:3,than:[0,1,3,6],them:0,ther:3,thi:[0,1,2,3,4,5,6,7],think:4,those:0,though:3,threadpoolexecutor:3,threshold:[0,2,3,4,5,7],throughput:3,thu:[3,4,7],tiger:4,time:[0,2,3,6],togeth:0,top:[0,2,3,5],toward:0,trade:6,train:4,tree:[0,5],tune:[3,4],tupl:[0,3],two:[0,1,3,4,6],type:[0,2,3,5,7],typeerror:0,under:0,underli:3,understand:5,uniform:7,union:[0,1,4,6],uniqu:[0,3],unique_name_6ac4fg:3,unit:0,univers:7,unlik:3,unseen:7,unsuit:7,until:0,unweight:7,updat:[0,1,3,4,5,6,7],usag:[0,1,2,6,7],use:[0,1,3,4,7],used:[0,1,2,3,4,5,6,7],useful:[3,5,6],uses:0,using:[0,1,2,3,4,5,6,7],usual:[3,7],utf8:[1,3,4,5,6],utf:[0,3],valu:[0,1,4,6],van:4,variabl:0,variou:3,vector:[0,7],venn:4,veri:[0,2,3,4,5,7],version:[0,1,3,4],wai:[3,4,5],want:[0,3,4,5],weak:0,weight:2,weightedminhash:[0,7],weightedminhashgener:[0,7],when:[0,3,4,7],where:0,whether:0,which:[0,3,4,5,7],whose:0,why:4,without:[0,5],wm1:7,wm2:7,wmg:7,wolf:4,won:[0,5],work:[0,4],you:[0,1,2,3,4,5,6,7],your:[3,4,6],yow:[0,3],zhu:[0,4],zip:[0,3]},titles:["API Documentation","HyperLogLog","datasketch: Big Data Looks Small","MinHash LSH","MinHash LSH Ensemble","MinHash LSH Forest","MinHash","Weighted MinHash"],titleterms:{api:0,asynchron:[0,3],big:2,connect:3,contain:4,data:2,datasketch:2,document:0,ensembl:[0,4],exist:3,forest:[0,5],hyperloglog:[0,1],instal:2,lean:0,look:2,lsh:[0,3,4,5],minhash:[0,3,4,5,6,7],scale:3,small:2,weight:[0,7]}})
